+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add props `children` on `Checkbox`.
+- Feature: Add `children` props to `Checkbox`.
 - Feature: Add new `HeaderTitles` component.
 - Fix: Update color `Checkbox` disabled.
 - Fix: Update color `RadioButton` disabled.